+++ POS/POS.Logic/CashDrawer/CashDrawerHistoryEventType.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace POS.Core.CashDrawer
+﻿namespace POS.Core.CashDrawer
 {
     public enum CashDrawerHistoryEventType
     {
         CashDrawerHistoryPrintSuccess = 2008,
         CashDrawerHistoryPrintFail = 2009
-
     }
 }